Separate out worker session properties and introduce JavaFeaturesConfig
