+++ presto-main/src/test/java/com/facebook/presto/sql/analyzer/TestFeaturesConfig.java
@@ -102,17 +102,6 @@ public class TestFeaturesConfig
                 .setAggregationPartitioningMergingStrategy(LEGACY)
                 .setSpillEnabled(false)
                 .setJoinSpillingEnabled(true)
-                .setAggregationSpillEnabled(true)
-                .setDistinctAggregationSpillEnabled(true)
-                .setDedupBasedDistinctAggregationSpillEnabled(false)
-                .setDistinctAggregationLargeBlockSpillEnabled(false)
-                .setDistinctAggregationLargeBlockSizeThreshold(DataSize.valueOf("50MB"))
-                .setOrderByAggregationSpillEnabled(true)
-                .setWindowSpillEnabled(true)
-                .setOrderBySpillEnabled(true)
-                .setTopNSpillEnabled(true)
-                .setAggregationOperatorUnspillMemoryLimit(DataSize.valueOf("4MB"))
-                .setTopNOperatorUnspillMemoryLimit(DataSize.valueOf("4MB"))
                 .setSpillerSpillPaths("")
                 .setSpillerThreads(4)
                 .setSpillMaxUsedSpaceThreshold(0.9)
@@ -323,17 +312,6 @@ public class TestFeaturesConfig
                 .put("optimizer.aggregation-partition-merging", "top_down")
                 .put("experimental.spill-enabled", "true")
                 .put("experimental.join-spill-enabled", "false")
-                .put("experimental.aggregation-spill-enabled", "false")
-                .put("experimental.distinct-aggregation-spill-enabled", "false")
-                .put("experimental.dedup-based-distinct-aggregation-spill-enabled", "true")
-                .put("experimental.distinct-aggregation-large-block-spill-enabled", "true")
-                .put("experimental.distinct-aggregation-large-block-size-threshold", "10MB")
-                .put("experimental.order-by-aggregation-spill-enabled", "false")
-                .put("experimental.window-spill-enabled", "false")
-                .put("experimental.order-by-spill-enabled", "false")
-                .put("experimental.topn-spill-enabled", "false")
-                .put("experimental.aggregation-operator-unspill-memory-limit", "100MB")
-                .put("experimental.topn-operator-unspill-memory-limit", "100MB")
                 .put("experimental.spiller-spill-path", "/tmp/custom/spill/path1,/tmp/custom/spill/path2")
                 .put("experimental.spiller-threads", "42")
                 .put("experimental.spiller-max-used-space-threshold", "0.8")
@@ -525,17 +503,6 @@ public class TestFeaturesConfig
                 .setPushAggregationThroughJoin(false)
                 .setSpillEnabled(true)
                 .setJoinSpillingEnabled(false)
-                .setAggregationSpillEnabled(false)
-                .setDistinctAggregationSpillEnabled(false)
-                .setDedupBasedDistinctAggregationSpillEnabled(true)
-                .setDistinctAggregationLargeBlockSpillEnabled(true)
-                .setDistinctAggregationLargeBlockSizeThreshold(DataSize.valueOf("10MB"))
-                .setOrderByAggregationSpillEnabled(false)
-                .setWindowSpillEnabled(false)
-                .setOrderBySpillEnabled(false)
-                .setTopNSpillEnabled(false)
-                .setAggregationOperatorUnspillMemoryLimit(DataSize.valueOf("100MB"))
-                .setTopNOperatorUnspillMemoryLimit(DataSize.valueOf("100MB"))
                 .setSpillerSpillPaths("/tmp/custom/spill/path1,/tmp/custom/spill/path2")
                 .setSpillerThreads(42)
                 .setSpillMaxUsedSpaceThreshold(0.8)
