+++ presto-spark-base/src/main/java/com/facebook/presto/spark/PrestoSparkSessionPropertyManagerProvider.java
@@ -20,6 +20,7 @@ import com.google.common.collect.Streams;
 import javax.inject.Inject;
 import javax.inject.Provider;
 
+import static com.facebook.presto.metadata.SessionPropertyManager.createTestingSessionPropertyManager;
 import static com.google.common.collect.ImmutableList.toImmutableList;
 import static java.util.Objects.requireNonNull;
 
@@ -39,9 +40,9 @@ public class PrestoSparkSessionPropertyManagerProvider
     @Override
     public SessionPropertyManager get()
     {
-        return new SessionPropertyManager(Streams.concat(
-                systemSessionProperties.getSessionProperties().stream(),
-                prestoSparkSessionProperties.getSessionProperties().stream())
+        return createTestingSessionPropertyManager(Streams.concat(
+                        systemSessionProperties.getSessionProperties().stream(),
+                        prestoSparkSessionProperties.getSessionProperties().stream())
                 .collect(toImmutableList()));
     }
 }
