+++ presto-spark-base/src/test/java/com/facebook/presto/spark/planner/TestPrestoSparkPhysicalResourceAllocationStrategy.java
@@ -16,7 +16,6 @@ package com.facebook.presto.spark.planner;
 import com.facebook.presto.Session;
 import com.facebook.presto.metadata.AbstractMockMetadata;
 import com.facebook.presto.metadata.Metadata;
-import com.facebook.presto.metadata.SessionPropertyManager;
 import com.facebook.presto.spark.PrestoSparkPhysicalResourceCalculator;
 import com.facebook.presto.spark.PrestoSparkSourceStatsCollector;
 import com.facebook.presto.spi.ColumnHandle;
@@ -41,6 +40,7 @@ import org.testng.annotations.Test;
 import java.util.List;
 
 import static com.facebook.presto.SystemSessionProperties.HASH_PARTITION_COUNT;
+import static com.facebook.presto.metadata.SessionPropertyManager.createTestingSessionPropertyManager;
 import static com.facebook.presto.spark.PrestoSparkSessionProperties.SPARK_AVERAGE_INPUT_DATA_SIZE_PER_EXECUTOR;
 import static com.facebook.presto.spark.PrestoSparkSessionProperties.SPARK_AVERAGE_INPUT_DATA_SIZE_PER_PARTITION;
 import static com.facebook.presto.spark.PrestoSparkSessionProperties.SPARK_EXECUTOR_ALLOCATION_STRATEGY_ENABLED;
@@ -85,12 +85,12 @@ public class TestPrestoSparkPhysicalResourceAllocationStrategy
             PropertyMetadata.integerProperty(HASH_PARTITION_COUNT, "HASH_PARTITION_COUNT", 150, false)
     };
     // system property with allocation based tuning enabled
-    private static final Session testSessionWithAllocation = testSessionBuilder(new SessionPropertyManager(
+    private static final Session testSessionWithAllocation = testSessionBuilder(createTestingSessionPropertyManager(
             new ImmutableList.Builder<PropertyMetadata<?>>().add(defaultPropertyMetadata).add(
                     PropertyMetadata.booleanProperty(SPARK_RESOURCE_ALLOCATION_STRATEGY_ENABLED, "SPARK_RESOURCE_ALLOCATION_STRATEGY_ENABLED", true, false)
             ).build())).build();
     // system property with allocation based tuning disabled
-    private static final Session testSessionWithoutAllocation = testSessionBuilder(new SessionPropertyManager(
+    private static final Session testSessionWithoutAllocation = testSessionBuilder(createTestingSessionPropertyManager(
             new ImmutableList.Builder<PropertyMetadata<?>>().add(defaultPropertyMetadata).add(
                     PropertyMetadata.booleanProperty(SPARK_RESOURCE_ALLOCATION_STRATEGY_ENABLED, "SPARK_RESOURCE_ALLOCATION_STRATEGY_ENABLED", false, false),
                     PropertyMetadata.booleanProperty(SPARK_HASH_PARTITION_COUNT_ALLOCATION_STRATEGY_ENABLED, "SPARK_HASH_PARTITION_COUNT_ALLOCATION_STRATEGY_ENABLED", false, false),
