+++ presto-tests/src/test/java/com/facebook/presto/tests/TestSpilledAggregationWithLargeBlockSpillingEnabled.java
@@ -21,8 +21,10 @@ import com.google.common.collect.ImmutableMap;
 
 import java.nio.file.Paths;
 
-import static com.facebook.presto.SystemSessionProperties.DISTINCT_AGGREGATION_LARGE_BLOCK_SIZE_THRESHOLD;
-import static com.facebook.presto.SystemSessionProperties.DISTINCT_AGGREGATION_LARGE_BLOCK_SPILL_ENABLED;
+import static com.facebook.presto.sessionpropertyproviders.JavaWorkerSessionPropertyProvider.AGGREGATION_OPERATOR_UNSPILL_MEMORY_LIMIT;
+import static com.facebook.presto.sessionpropertyproviders.JavaWorkerSessionPropertyProvider.DEDUP_BASED_DISTINCT_AGGREGATION_SPILL_ENABLED;
+import static com.facebook.presto.sessionpropertyproviders.JavaWorkerSessionPropertyProvider.DISTINCT_AGGREGATION_LARGE_BLOCK_SIZE_THRESHOLD;
+import static com.facebook.presto.sessionpropertyproviders.JavaWorkerSessionPropertyProvider.DISTINCT_AGGREGATION_LARGE_BLOCK_SPILL_ENABLED;
 import static com.facebook.presto.testing.TestingSession.testSessionBuilder;
 import static com.facebook.presto.tpch.TpchMetadata.TINY_SCHEMA_NAME;
 
@@ -37,9 +39,9 @@ public class TestSpilledAggregationWithLargeBlockSpillingEnabled
                 .setCatalog("tpch")
                 .setSchema(TINY_SCHEMA_NAME)
                 .setSystemProperty(SystemSessionProperties.TASK_CONCURRENCY, "2")
-                .setSystemProperty(SystemSessionProperties.AGGREGATION_OPERATOR_UNSPILL_MEMORY_LIMIT, "128kB")
+                .setSystemProperty(AGGREGATION_OPERATOR_UNSPILL_MEMORY_LIMIT, "128kB")
                 .setSystemProperty(SystemSessionProperties.USE_MARK_DISTINCT, "false")
-                .setSystemProperty(SystemSessionProperties.DEDUP_BASED_DISTINCT_AGGREGATION_SPILL_ENABLED, "true")
+                .setSystemProperty(DEDUP_BASED_DISTINCT_AGGREGATION_SPILL_ENABLED, "true")
                 // Enable Large block spilling
                 .setSystemProperty(DISTINCT_AGGREGATION_LARGE_BLOCK_SPILL_ENABLED, "true")
                 .setSystemProperty(DISTINCT_AGGREGATION_LARGE_BLOCK_SIZE_THRESHOLD, "150B")
